separate function and interface in vizualisation
