+++ ed_sensor_integration/CMakeLists.txt
@@ -113,8 +113,8 @@ add_dependencies(ed_image_saver ${catkin_EXPORTED_TARGETS})
 add_executable(ed_segmenter tools/segmenter.cpp)
 target_link_libraries(ed_segmenter ed_kinect)
 
-add_executable(ed_fitter tools/fitter_viz.cpp)
-target_link_libraries(ed_fitter ed_kinect)
+add_executable(ed_fitter_data tools/fitter_viz_data.cpp)
+target_link_libraries(ed_fitter_data ed_kinect)
 
 # ------------------------------------------------------------------------------------------------
 #                                           TESTS
